Cleanups, some progress - now having trouble with fat static libraries

+++ src/libc/darwin/init_cpu_capabilities.c
@@ -25,7 +25,7 @@
 #include <machine/cpu_capabilities.h>
 #undef	__APPLE_API_PRIVATE
 
-extern int _get_cpu_capabilities(void);
+extern uint64_t _get_cpu_capabilities(void);
 
 int _cpu_capabilities = 0;
 int _cpu_has_altivec = 0;		// DEPRECATED: use _cpu_capabilities instead
